Add recovered code for approx, (reimplemented in master, as I wasn't able to access the cluster for some time)

+++ python/stop.sh
@@ -1,7 +1,11 @@
 ext=${1:-"*"}
 
-echo Stopping containers ...
-docker container stop $(docker container ls -q --filter name="mmatusovic_python_$ext")
+if [[ ! $(docker container ls -q --filter name="mmatusovic_python_$ext") ]]; then
+    echo "No containers matching \"mmatusovic_python_$ext\" running."
+else
+    echo Stopping containers ...
+    docker container stop $(docker container ls -q --filter name="mmatusovic_python_$ext")
 
-echo Removing containers ...
-docker container rm $(docker container ls -a -q --filter name="mmatusovic_python_$ext")
+    echo Removing containers ...
+    docker container rm $(docker container ls -a -q --filter name="mmatusovic_python_$ext")
+fi
